LibTest: Move Crash testing facility from crash.cpp to LibTest
