+++ Userland/Libraries/LibTest/CMakeLists.txt
@@ -2,6 +2,7 @@ serenity_install_sources("Userland/Libraries/LibTest")
 
 set(SOURCES
     TestSuite.cpp
+    CrashTest.cpp
 )
 
 serenity_lib(LibTest test)
