+++ Userland/Utilities/CMakeLists.txt
@@ -22,6 +22,7 @@ target_link_libraries(checksum LibCrypto)
 target_link_libraries(chres LibGUI)
 target_link_libraries(cksum LibCrypto)
 target_link_libraries(copy LibGUI)
+target_link_libraries(crash LibTest)
 target_link_libraries(disasm LibX86)
 target_link_libraries(expr LibRegex)
 target_link_libraries(functrace LibDebug LibX86)
